module-code: separate exports and imports from the same file
This avoids adding cyclic dependency constraints on the module loader
to test the actual subject under test.
