+++ test/language/module-code/eval-export-dflt-cls-anon.js
@@ -26,8 +26,7 @@ info: |
 flags: [module]
 ---*/
 
-export default class { valueOf() { return 45; } }
-import C from './eval-export-dflt-cls-anon.js';
+import C from './eval-export-dflt-cls-anon_FIXTURE.js';
 
 assert.sameValue(new C().valueOf(), 45, 'binding initialized');
 assert.sameValue(C.name, 'default', 'correct name is assigned');
