+++ test/language/module-code/eval-export-dflt-cls-name-meth.js
@@ -26,8 +26,7 @@ info: |
 flags: [module]
 ---*/
 
-export default class { static name() { return 'name method'; } }
-import C from './eval-export-dflt-cls-name-meth.js';
+import C from './eval-export-dflt-cls-name-meth_FIXTURE.js';
 
 assert.sameValue(
   C.name(), 'name method', '`name` property is not over-written'
