+++ test/language/module-code/eval-export-dflt-cls-name-meth_FIXTURE.js
@@ -1,0 +1,4 @@
+// Copyright (C) 2016 the V8 project authors. All rights reserved.
+// This code is governed by the BSD license found in the LICENSE file.
+
+export default class { static name() { return 'name method'; } }
