+++ test/language/module-code/eval-export-dflt-expr-cls-anon.js
@@ -25,8 +25,7 @@ info: |
 flags: [module]
 ---*/
 
-export default (class { valueOf() { return 45; } });
-import C from './eval-export-dflt-expr-cls-anon.js';
+import C from './eval-export-dflt-expr-cls-anon_FIXTURE.js';
 
 assert.sameValue(new C().valueOf(), 45, 'binding initialized');
 assert.sameValue(C.name, 'default', 'correct name is assigned');
