+++ test/language/module-code/eval-export-dflt-expr-cls-name-meth.js
@@ -27,8 +27,7 @@ info: |
 flags: [module]
 ---*/
 
-export default (class { static name() { return 'name method'; } });
-import C from './eval-export-dflt-expr-cls-name-meth.js';
+import C from './eval-export-dflt-expr-cls-name-meth_FIXTURE.js';
 
 assert.sameValue(
   C.name(), 'name method', '`name` property is not over-written'
